Move StandardSum tests to tests.rs; simplify docs
